requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
aioredis==1.3.1
aiorpcX==0.22.1
amqp==5.0.6
anyio==3.3.0
appdirs==1.4.4
APScheduler==3.6.3
arrow==1.1.1
asgiref==3.4.1
asn1crypto==1.4.0
aspy.yaml==1.3.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrdict==2.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
base58==2.1.0
base58check==1.0.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
bech32==1.2.0
behave==1.2.6
behave-django==1.4.0
billiard==3.6.4.0
bit==0.7.2
bitarray==1.2.2
bitcoin==1.1.42
cached-property==1.5.2
cachetools==4.2.2
cashaddress==1.0.6
cchardet==2.1.7
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coinaddr==1.0.1
coincurve==15.0.1
constantly==15.1.0
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
cosmospy==6.0.0
coverage==5.5
crypto==1.4.1
cryptography==3.4.8
cryptos @ https://github.com/primal100/pybitcointools/archive/master.zip
cytoolz==0.11.0
daphne==3.0.2
dateparser==1.0.0
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.7
django-admin-rangefilter==0.9.0
django-admin-tools==0.9.2
django-allauth==0.45.0
django-bulk-update==2.2.0
django-celery-results==2.2.0
django-chunked-iterator==0.6.1
django-cors-headers==3.8.0
django-countries==7.2.1
django-environ==0.9.0
django-filter==2.4.0
django-import-export==2.5.0
django-ipware==4.0.0
django-maintenance-mode==0.16.0
django-model-utils==4.1.1
django-modeltranslation==0.17.3
django-otp==1.0.6
django-pkgconf==0.4.0
django-qsstats-magic==1.1.0
django-redis==5.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-tabbed-admin==1.0.4
django-user-agents==0.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dnspython==2.1.0
docker-py==1.10.6
docker-pycreds==0.4.0
drf-spectacular==0.23.1
ecdsa==0.17.0
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
ethereum-input-decoder==0.2.2
fabric==2.6.0
filelock==3.0.12
flake8==3.9.2
flower==1.0.0
funcsigs==1.0.2
funcy==1.16
future==0.18.2
fuzzywuzzy==0.18.0
graypy==2.1.0
gunicorn==20.1.0
h11==0.12.0
hdwallets==0.1.2
hexbytes==0.2.2
hiredis==2.0.0
httpcore==0.13.7
httpx==0.18.2
humanize==3.11.0
hyperlink==21.0.0
identify==2.2.13
idna==3.2
idna-ssl==1.1.0
immutables==0.16
importlib-metadata==4.8.1
importlib-resources==5.2.2
incremental==21.3.0
iniconfig==1.1.1
invoke==1.6.0
ipfshttpclient==0.8.0a2
ipython==7.27.0
ipython-genutils==0.2.0
isodate==0.6.0
itypes==1.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.1
jsonfield==3.1.0
jsonrpcclient==4.0.0
jsonrpclib-pelix==0.4.2
jsonrpcserver==5.0.3
jsonschema==3.2.0
keyring==23.1.0
keyrings.alt==4.1.0
kombu==5.1.0
lru-dict==1.1.7
lxml==4.6.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mnemonic==0.20
model-mommy==2.0.0
more-itertools==8.8.0
msgpack==1.0.2
multiaddr==0.0.9
multidict==5.1.0
Naked==0.1.31
netaddr==0.8.0
nodeenv==1.6.0
numpy==1.21.2
oauthlib==3.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
OSlash==0.6.3
packaging==21.0
paramiko==2.7.2
parse==1.19.0
parse-type==0.5.2
parsimonious==0.8.1
parso==0.8.2
path==16.2.0
path.py==12.5.0
pathlib2==2.3.6
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
pluggy==1.0.0
pre-commit==2.15.0
prettytable==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
pyflakes==2.3.1
Pygments==2.10.0
PyHamcrest==2.0.2
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyotp==2.6.0
pyparsing==2.4.7
pyrsistent==0.18.0
pysha3==1.0.2
PySocks==1.7.1
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
python-binance==1.0.12
python-bitcoinrpc==1.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-memcached==1.59
python-ripple-lib==1.0.11
python-socks==1.2.4
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2021.1
pywallet @ https://github.com/polyxexchange/pywallet/archive/master.zip
pyxdg==0.27
PyYAML==5.4.1
QDarkStyle==3.0.2
qrcode==7.3
QtPy==1.10.0
redis==3.5.3
regex==2021.8.28
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rlp==2.0.1
SecretStorage==3.3.1
sentry-sdk==1.3.1
service-identity==21.1.0
sha256==0.1
shellescape==3.8.1
simplegeneric==0.8.1
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
tabulate==0.8.9
tblib==1.7.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
twilio==6.50.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzlocal==3.0
ua-parser==0.10.0
ujson==4.1.0
uritemplate==3.0.1
urllib3==1.26.6
user-agents==2.2.0
varint==1.0.2
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
web3==5.23.1
websocket-client==1.2.1
websockets==9.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
zeep==4.1.0
zipp==3.5.0
zope.interface==5.4.0
